Add authorize middleware

+++ routes.php
@@ -16,19 +16,19 @@
 
 $router->get("/", "HomeController@index");
 $router->get("/listings", "ListingsController@index");
-$router->get("/listings/create", "ListingsController@getCreate");
-$router->get("/listings/edit/{id}", "ListingsController@edit");
+$router->get("/listings/create", "ListingsController@getCreate", ["auth"]);
+$router->get("/listings/edit/{id}", "ListingsController@edit", ["auth"]);
 $router->get("/listings/{id}", "ListingsController@show");
 
-$router->post("/listings", "ListingsController@store");
-$router->put("/listings/{id}", "ListingsController@update");
+$router->post("/listings", "ListingsController@store", ["auth"]);
+$router->put("/listings/{id}", "ListingsController@update", ["auth"]);
 
-$router->delete("/listings/{id}", "ListingsController@destroy");
+$router->delete("/listings/{id}", "ListingsController@destroy", ["auth"]);
 
 // USER ROUTES
-$router->get("/auth/register", "UserController@create");
-$router->get("/auth/login", "UserController@login");
+$router->get("/auth/register", "UserController@create", ["guest"]);
+$router->get("/auth/login", "UserController@login", ["guest"]);
 
-$router->post("/auth/register", "UserController@store");
-$router->get("/auth/logout", "UserController@logout");
-$router->post("/auth/login", "UserController@authenticate");
+$router->post("/auth/register", "UserController@store", ["guest"]);
+$router->get("/auth/logout", "UserController@logout", ["auth"]);
+$router->post("/auth/login", "UserController@authenticate", ["guest"]);
